+++ modules/load-balancer/main.tf
@@ -22,24 +22,30 @@ resource "aws_security_group" "lb_sg" {
 }
 
 resource "aws_security_group_rule" "http" {
+  count = var.allow_http ? 1 : 0
+
   type              = "ingress"
   from_port         = 80
   to_port           = 80
   protocol          = "tcp"
-  cidr_blocks       = ["0.0.0.0/0"]
+  cidr_blocks       = var.cidr_blocks
   security_group_id = aws_security_group.lb_sg.id
 }
 
 resource "aws_security_group_rule" "https" {
+  count = var.allow_https ? 1 : 0
+
   type              = "ingress"
   from_port         = 443
   to_port           = 443
   protocol          = "tcp"
-  cidr_blocks       = ["0.0.0.0/0"]
+  cidr_blocks       = var.cidr_blocks
   security_group_id = aws_security_group.lb_sg.id
 }
 
 resource "aws_security_group_rule" "allow_all_outbound_lb" {
+  count = var.allow_all_outbound ? 1 : 0
+
   type              = "egress"
   from_port         = 0
   to_port           = 0